fix: move instatus to notification workspace and register it as a notification channel (#1056)
