+++ packages/notification/channel/instatus/database.ts
@@ -1,4 +1,6 @@
-import { db } from '../../../components/logger/history'
+import path from 'path'
+import sqlite3 from 'sqlite3'
+import { open, type Database } from 'sqlite'
 
 type Incident = {
   id: string
@@ -15,6 +17,9 @@ type FindIncidentResponse = {
   // eslint-disable-next-line camelcase
   incident_id: string
 }
+
+const dbPath = path.resolve(process.cwd(), 'monika-logs.db')
+let db: Database<sqlite3.Database, sqlite3.Statement>
 
 export async function insertIncident({
   status,
@@ -33,7 +38,7 @@ export async function insertIncident({
   ) VALUES (?, ?, ?, ?, ?, ?);`
   const sqlParams = [status, url, probeID, incidentID, dateNow, dateNow]
 
-  await db.run(sqlStatement, sqlParams)
+  await (await getDatabase()).run(sqlStatement, sqlParams)
 }
 
 export async function updateIncident({
@@ -45,7 +50,7 @@ export async function updateIncident({
    WHERE incident_id = ?`
   const sqlParams = [status, dateNow, incidentID]
 
-  await db.run(sqlStatement, sqlParams)
+  await (await getDatabase()).run(sqlStatement, sqlParams)
 }
 
 export async function findIncident({
@@ -54,9 +59,25 @@ export async function findIncident({
   url,
 }: FindIncident): Promise<FindIncidentResponse | undefined> {
   const sqlStatement = `SELECT incident_id FROM instatus_page_incidents 
-    WHERE status = ? AND url = ? AND probe_id = ?`
+    WHERE status = ? AND url = ? AND probe_id = ? LIMIT 1`
   const sqlParams = [status, url, probeID]
-  const incident = await db.get<FindIncidentResponse>(sqlStatement, sqlParams)
+  const incident = await (
+    await getDatabase()
+  ).get<FindIncidentResponse>(sqlStatement, sqlParams)
 
   return incident
+}
+
+async function getDatabase(): Promise<
+  Database<sqlite3.Database, sqlite3.Statement>
+> {
+  if (!db) {
+    db = await open({
+      filename: dbPath,
+      mode: sqlite3.OPEN_READWRITE | sqlite3.OPEN_CREATE,
+      driver: sqlite3.Database,
+    })
+  }
+
+  return db
 }
