+++ packages/notification/package.json
@@ -36,6 +36,8 @@
     "mailgun.js": "^3.5.4",
     "msw": "^0.36.4",
     "nodemailer": "^6.5.0",
-    "pino": "6.11.2"
+    "pino": "6.11.2",
+    "sqlite": "4.2.1",
+    "sqlite3": "5.1.6"
   }
 }
